fix: address PR review issues for subagent orchestration
Fixes:
- [critical] ConcurrencyPool.abortAll() now cleans up callCounts (was leaking
  per-call counts for aborted slots). Track callGroup in SlotInfo.
- [critical] SubagentEntry type divergence — electron/subagent/types.ts now
  re-exports IPC-crossing types from src/types/subagent.ts (single source).
- [critical] agent.ts exceeded 500 LOC — extracted subscribeToSession() to
  agent-subscription.ts (507 → 369 lines).
- [warning] abort IPC handler now calls SubagentManager.abortOne() which
  signals the AbortController via the pool, not just tracker.abort().
  Added ConcurrencyPool.abortOne() and SubagentManager.abortOne().
- [warning] writeAgent used dynamic import('fs/promises') for rename —
  moved to static import at top of file per AGENTS.md convention.
- [warning] shared-infra.ts accessed private deps via bracket notation —
  added public isInitialized getter on SubagentManager.
- [security] Added path traversal validation (VALID_AGENT_NAME regex) to
  readAgent, writeAgent, and deleteAgent IPC handlers.
